flaskk/requirements.txt
Flask
Flask-Cors
Pillow 
requests
python-dotenv
pybase64
google-generativeai
